Tutorial 6/requirements.txt
flask-3.0.0 
Jinja2-3.1.2 
MarkupSafe-2.1.3 
Werkzeug-3.0.1 
blinker-1.7.0 
click-8.1.7 
colorama-0.4.6 
itsdangerous-2.1.2
numpy-1.26.1 
python-dateutil-2.8.2 
pytz-2023.3.post1 
pipsix-1.16.0
joblib-1.3.2 
scipy-1.11.3 
threadpoolctl-3.2.0
numpy-1.26.1 
opencv-contrib-python-4.8.1.78